Add recipe descriptor to toplevel class Refactor declarative recipe tests into one class with nested subclasses Combine cache related tests into a single test (which gets rid of the duplicate @DocumentExample)
